fix: merge history pages
