+++ backend/core/tests.py
@@ -504,191 +504,3 @@ class RouteViewSetTests(APITestCase):
         self.assertEqual(response.data['upvotes_count'], 2)
         self.assertEqual(response.data['downvotes_count'], 0)
         self.assertEqual(response.data['user_vote'], 'upvote')
-
-class RouteHistoryTests(APITestCase):
-    """Tests for the route history functionality"""
-    
-    def setUp(self):
-        """Create test data for history tests"""
-        # Create test users
-        self.user = User.objects.create_user(
-            username='historyuser',
-            email='history@example.com',
-            password='historypass123'
-        )
-        
-        self.other_user = User.objects.create_user(
-            username='otherhistoryuser',
-            email='otherhistory@example.com',
-            password='otherpass123'
-        )
-        
-        # Get JWT token for main test user
-        refresh = RefreshToken.for_user(self.user)
-        self.access_token = str(refresh.access_token)
-        self.client.credentials(HTTP_AUTHORIZATION=f'Bearer {self.access_token}')
-        
-        # Create user details for both users (should normally be created by signal)
-        self.user_details = UserDetails.objects.create(user=self.user)
-        self.other_user_details = UserDetails.objects.create(user=self.other_user)
-        
-        # Create test routes
-        self.route1 = Route.objects.create(
-            title="History Test Route 1",
-            description="First route for history testing",
-            starting_location="Start Point 1",
-            ending_location="End Point 1",
-            coordinates=[[40.7128, -74.0060], [40.7129, -74.0061]],
-            user=self.user
-        )
-        
-        self.route2 = Route.objects.create(
-            title="History Test Route 2",
-            description="Second route for history testing",
-            starting_location="Start Point 2",
-            ending_location="End Point 2",
-            coordinates=[[34.0522, -118.2437], [34.0523, -118.2438]],
-            user=self.user
-        )
-        
-        self.route3 = Route.objects.create(
-            title="History Test Route 3",
-            description="Third route for history testing",
-            starting_location="Start Point 3",
-            ending_location="End Point 3",
-            coordinates=[[41.8781, -87.6298], [41.8782, -87.6299]],
-            user=self.other_user
-        )
-    
-    def test_add_route_to_history(self):
-        """Test adding a route to the user's history"""
-        url = reverse('route-detail', args=[self.route1.id]) + 'add_to_history/'
-        response = self.client.post(url)
-        
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertIn('message', response.data)
-        
-        # Check if route was added to user's history
-        self.user_details.refresh_from_db()
-        history = self.user_details.route_history
-        self.assertEqual(len(history), 1)
-        self.assertEqual(history[0]['route_id'], self.route1.id)
-    
-    def test_get_route_history(self):
-        """Test retrieving the user's route history"""
-        # Add some routes to history
-        self.user_details.add_to_history(self.route1.id, self.route1.title)
-        self.user_details.add_to_history(self.route2.id, self.route2.title)
-        
-        url = reverse('user-details-route-history')
-        response = self.client.get(url)
-        
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(len(response.data), 2)
-        
-        # Latest added route should be first
-        self.assertEqual(response.data[0]['route_id'], self.route2.id)
-        self.assertEqual(response.data[1]['route_id'], self.route1.id)
-    
-    def test_get_limited_route_history(self):
-        """Test retrieving a limited number of history entries"""
-        # Add routes to history
-        for i in range(5):
-            self.user_details.add_to_history(
-                self.route1.id, 
-                f"View {i} of route 1"
-            )
-        
-        url = reverse('user-details-route-history') + '?limit=3'
-        response = self.client.get(url)
-        
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(len(response.data), 3)  # Should be limited to 3
-    
-    def test_clear_route_history(self):
-        """Test clearing the user's route history"""
-        # Add some routes to history first
-        self.user_details.add_to_history(self.route1.id, self.route1.title)
-        self.user_details.add_to_history(self.route2.id, self.route2.title)
-        
-        url = reverse('user-details-clear-history')
-        response = self.client.delete(url)
-        
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        
-        # Check that history is now empty
-        self.user_details.refresh_from_db()
-        self.assertEqual(len(self.user_details.route_history), 0)
-    
-    def test_user_specific_history(self):
-        """Test that history is specific to each user"""
-        # Add routes to first user's history
-        self.user_details.add_to_history(self.route1.id, self.route1.title)
-        
-        # Add different routes to second user's history
-        # First switch to the other user
-        refresh = RefreshToken.for_user(self.other_user)
-        self.client.credentials(HTTP_AUTHORIZATION=f'Bearer {str(refresh.access_token)}')
-        
-        # Add route to other user's history
-        other_url = reverse('route-detail', args=[self.route3.id]) + 'add_to_history/'
-        self.client.post(other_url)
-        
-        # Get other user's history
-        history_url = reverse('user-details-route-history')
-        response = self.client.get(history_url)
-        
-        # Verify other user only sees their history
-        self.assertEqual(len(response.data), 1)
-        self.assertEqual(response.data[0]['route_id'], self.route3.id)
-        
-        # Switch back to first user
-        self.client.credentials(HTTP_AUTHORIZATION=f'Bearer {self.access_token}')
-        
-        # Get first user's history
-        response = self.client.get(history_url)
-        
-        # Verify first user only sees their history
-        self.assertEqual(len(response.data), 1)
-        self.assertEqual(response.data[0]['route_id'], self.route1.id)
-    
-    def test_duplicate_entries_in_history(self):
-        """Test that adding the same route multiple times updates instead of duplicating"""
-        # Add the same route twice
-        url = reverse('route-detail', args=[self.route1.id]) + 'add_to_history/'
-        self.client.post(url)
-        self.client.post(url)
-        
-        # Get history
-        history_url = reverse('user-details-route-history')
-        response = self.client.get(history_url)
-        
-        # Should only have one entry
-        self.assertEqual(len(response.data), 1)
-    
-    def test_history_with_deleted_route(self):
-        """Test handling deleted routes in history"""
-        # Add route to history
-        self.user_details.add_to_history(self.route1.id, self.route1.title)
-        
-        # Delete the route
-        self.route1.delete()
-        
-        # Get history
-        url = reverse('user-details-route-history')
-        response = self.client.get(url)
-        
-        # Should still return history with the entry
-        self.assertEqual(len(response.data), 1)
-        self.assertEqual(response.data[0]['route_id'], self.route1.id)
-        # No additional route details should be present since route is deleted
-    
-    def test_unauthenticated_access_to_history(self):
-        """Test that unauthenticated users cannot access history"""
-        # Remove credentials
-        self.client.credentials()
-        
-        url = reverse('user-details-route-history')
-        response = self.client.get(url)
-        
-        self.assertEqual(response.status_code, status.HTTP_401_UNAUTHORIZED)
